Use %env_asan_opts= substitution instead of 'env ASAN_OPTIONS=$ASAN_OPTIONS'
Summary:
The lit internal shell is used by default on Windows, and it does not support
bash variable expansion. Because bash variable expansion interacts with
tokenization, it is prohibitively difficult to make the existing lit shell do
general shell variable expansion.

The most common use of shell variables in the asan tests is to add options to
the default set of options set by lit.cfg. We can avoid the need for variable
expansion with a substitution that expands to 'env ASAN_OPTIONS=<defaults:>'.

This has the side benefit of shortening the RUN lines, so it seemed better than
implementing limited variable expansion in lit.

Reviewers: samsonov, filcab

Subscribers: llvm-commits

Differential Revision: http://reviews.llvm.org/D11982